pyproject.toml
[project]
name = "ask"
version = "0.8.1"
description = "Lookup CLI commands easily"
readme = "README.md"
dependencies = [
    "openai>=1.72.0",
    "pydantic>=2.10.6",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.0.1",
    "questionary>=2.1.0",
    "rich>=13.9.4"
]
requires-python = ">=3.9"
urls = { Repository = "https://github.com/dtnewman/ask" }

[project.scripts]
ask = "ask.main:app"

[project.optional-dependencies]
dev = [
    "ruff>=0.11.2",
    "isort>=5.13.2",
    "pytest>=7.0.0",
]
azure = [
    "azure-identity>=1.20.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["ask*"]

[tool.ruff]
line-length = 120
# this ensures imports are properly sorted
lint.select = ["I001"]

[tool.isort]
profile = "black"
line_length = 120
